添加角色和菜单相关的数据操作

+++ server/src/main/java/com/danny/lvsen/mapper/BillMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface BillMapper {
-    long countByExample(BillExample example);
+    int countByExample(BillExample example);
 
     int deleteByExample(BillExample example);
