+++ server/src/main/java/com/danny/lvsen/mapper/GoodsMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface GoodsMapper {
-    long countByExample(GoodsExample example);
+    int countByExample(GoodsExample example);
 
     int deleteByExample(GoodsExample example);
