+++ server/src/main/java/com/danny/lvsen/mapper/OrderMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface OrderMapper {
-    long countByExample(OrderExample example);
+    int countByExample(OrderExample example);
 
     int deleteByExample(OrderExample example);
